A groovebox using arduino
digitalgroovebox is an attempt to create a usable groovebox using an
arduino uno, an OLED screen, 6 buttons and two rotary encoders. In order
to do that, we build upon other projects, namely Mozzi, Adafruit_SSD1306
and of course the illutron b wavetable synth engine. Planned features.

4 operating modes STOP (realtime jam), PLAY (playback recorded samples,
RECORD (record a pattern in realtime) and STEP_RECORD (record a pattern
using a simple step sequencer approach).
The groovebox features 6 buttons. Of which the 4 are for sound
triggering (used as drum/sound pads) one shift button that enables
secondary functions of the 4 drum pads, and another mode button used
together with the pads in order to change operating mode. The two rotary
encoders can control various engine parameters. This part is not yet
complete.